+++ pkg/interface/src/logic/lib/reactNative.ts
@@ -1,0 +1,5 @@
+export const postReactNativeMessage = (data: any) => {
+  if ((window as any).ReactNativeWebView?.postMessage) {
+    (window as any).ReactNativeWebView.postMessage(JSON.stringify(data));
+  }
+};
